requirements.txt
# requirements.txt

beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
deep-translator==1.11.4
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
gTTS==2.5.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pillow==11.0.0
proglog==0.1.10
pydub==0.25.1
regex==2024.9.11
requests==2.32.3
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
